k9_aif_abb/requirements.txt
pydantic>=2.0
PyYAML>=6.0
requests>=2.30
python-dotenv>=1.0
rich>=13.0
kafka-python>=2.0.2

crewai>=0.170
litellm>=1.0
mcp>=1.0

fastapi>=0.110
uvicorn>=0.30
websockets>=12.0
sse-starlette>=2.0

prometheus-client>=0.20
paho-mqtt>=2.0

pandas>=2.0
openpyxl>=3.1

pypdf>=5.0
pdfplumber>=0.11
python-docx>=1.0

chromadb>=0.5
qdrant-client>=1.0

openai>=1.0
anthropic>=0.30
cohere>=5.0
